Refactor `TTSController`
