+++ Sources/Navigator/TTS/TTSController.swift
@@ -4,7 +4,6 @@
 //  available in the top-level LICENSE file of the project.
 //
 
-import AVFoundation
 import Foundation
 import R2Shared
 
@@ -50,7 +49,7 @@ public class TTSController: Loggable, TTSEngineDelegate {
         }
     }
 
-    public class func canSpeak(_ publication: Publication) -> Bool {
+    public static func canPlay(_ publication: Publication) -> Bool {
         publication.isContentIterable
     }
 
@@ -83,10 +82,9 @@ public class TTSController: Loggable, TTSEngineDelegate {
 
     public var isPlaying: Bool = false {
         didSet {
+            precondition(Thread.isMainThread, "TTSController.isPlaying must be mutated from the main thread")
             if oldValue != isPlaying {
-                DispatchQueue.main.async { [self] in
-                    delegate?.ttsController(self, playingDidChange: isPlaying)
-                }
+                delegate?.ttsController(self, playingDidChange: isPlaying)
             }
         }
     }
@@ -97,6 +95,12 @@ public class TTSController: Loggable, TTSEngineDelegate {
         } else {
             play(from: start)
         }
+    }
+
+    public func pause() {
+        precondition(Thread.isMainThread, "TTSController.pause() must be called from the main thread")
+        isPlaying = false
+        engine.stop()
     }
 
     public func play(from start: Locator? = nil) {
@@ -117,32 +121,40 @@ public class TTSController: Loggable, TTSEngineDelegate {
         }
     }
 
-    private func play(_ utterance: TTSUtterance) {
-        DispatchQueue.main.async { [self] in
-            delegate?.ttsController(self, willStartSpeaking: utterance)
-            isPlaying = true
-            engine.speak(utterance)
-        }
-    }
-
-    public func pause() {
-        isPlaying = false
-        engine.stop()
+    public func previous() {
+        playNextUtterance(direction: .backward)
     }
 
     public func next() {
+        playNextUtterance(direction: .forward)
+    }
+
+    private enum Direction {
+        case forward, backward
+    }
+
+    private var contentIterator: ContentIterator? {
+        willSet { contentIterator?.close() }
+    }
+
+    private var speakingUtteranceIndex: Int?
+    private var utterances: [TTSUtterance] = []
+
+    private var currentUtterance: TTSUtterance? {
+        speakingUtteranceIndex.map { utterances[$0] }
+    }
+
+    private func playNextUtterance(direction: Direction) {
         queue.async { [self] in
             do {
-                guard let utterance = try nextUtterance(direction: .forward) else {
-                    isPlaying = false
-                    return
+                let utterance = try nextUtterance(direction: direction)
+                DispatchQueue.main.async {
+                    if let utterance = utterance {
+                        play(utterance)
+                    } else {
+                        isPlaying = false
+                    }
                 }
-                if !utterance.text.contains(where: { $0.isLetter || $0.isNumber }) {
-                    next()
-                    return
-                }
-                play(utterance)
-
             } catch {
                 DispatchQueue.main.async {
                     delegate?.ttsController(self, didReceiveError: error)
@@ -151,21 +163,12 @@ public class TTSController: Loggable, TTSEngineDelegate {
         }
     }
 
-    // MARK: - Utterances
+    private func play(_ utterance: TTSUtterance) {
+        precondition(Thread.isMainThread, "TTSController.play() must be called from the main thread")
 
-    private var contentIterator: ContentIterator? {
-        willSet { contentIterator?.close() }
-    }
-
-    private enum Direction {
-        case forward, backward
-    }
-
-    private var speakingUtteranceIndex: Int?
-    private var utterances: [TTSUtterance] = []
-
-    private var currentUtterance: TTSUtterance? {
-        speakingUtteranceIndex.map { utterances[$0] }
+        delegate?.ttsController(self, willStartSpeaking: utterance)
+        isPlaying = true
+        engine.speak(utterance)
     }
 
     private func nextUtterance(direction: Direction) throws -> TTSUtterance? {
@@ -227,10 +230,10 @@ public class TTSController: Loggable, TTSEngineDelegate {
             guard let description = description, !description.isEmpty else {
                 return []
             }
-            return [utterance(text: description, locator: content.locator)]
+            return Array(ofNotNil: utterance(text: description, locator: content.locator))
 
         case .text(spans: let spans, style: _):
-            return spans.enumerated().map { offset, span in
+            return spans.enumerated().compactMap { offset, span in
                 utterance(
                     text: span.text,
                     locator: span.locator,
@@ -241,8 +244,11 @@ public class TTSController: Loggable, TTSEngineDelegate {
         }
     }
 
-    private func utterance(text: String, locator: Locator, language: String? = nil, postDelay: TimeInterval = 0) -> TTSUtterance {
-        TTSUtterance(
+    private func utterance(text: String, locator: Locator, language: String? = nil, postDelay: TimeInterval = 0) -> TTSUtterance? {
+        guard text.contains(where: { $0.isLetter || $0.isNumber }) else {
+            return nil
+        }
+        return TTSUtterance(
             text: text,
             locator: locator,
             language: language ?? defaultLanguage,
@@ -252,10 +258,10 @@ public class TTSController: Loggable, TTSEngineDelegate {
         )
     }
 
-    private var defaultLanguage: String {
+    private var defaultLanguage: String? {
         config.defaultLanguage
             ?? publication.metadata.languages.first
-            ?? AVSpeechSynthesisVoice.currentLanguageCode()
+            ?? engine.defaultLanguage
     }
 
     // MARK: - TTSEngineDelegate
@@ -269,106 +275,6 @@ public class TTSController: Loggable, TTSEngineDelegate {
     public func ttsEngine(_ engine: TTSEngine, willSpeakRangeAt locator: Locator, of utterance: TTSUtterance) {
         DispatchQueue.main.async { [self] in
             delegate?.ttsController(self, willSpeakRangeAt: locator, of: utterance)
-        }
-    }
-}
-
-public protocol TTSEngineDelegate: AnyObject {
-    func ttsEngine(_ engine: TTSEngine, willSpeakRangeAt locator: Locator, of utterance: TTSUtterance)
-    func ttsEngine(_ engine: TTSEngine, didFinish utterance: TTSUtterance)
-}
-
-public protocol TTSEngine: AnyObject {
-    var defaultRate: Double? { get }
-    var defaultPitch: Double? { get }
-    var delegate: TTSEngineDelegate? { get set }
-
-    func speak(_ utterance: TTSUtterance)
-    func stop()
-}
-
-public class AVTTSEngine: NSObject, TTSEngine, AVSpeechSynthesizerDelegate, Loggable {
-
-    public lazy var defaultRate: Double? =
-        avRateRange.percentageForValue(Double(AVSpeechUtteranceDefaultSpeechRate))
-
-    /// Range of valid values for an AVUtterance rate.
-    ///
-    /// > The speech rate is a decimal representation within the range of `AVSpeechUtteranceMinimumSpeechRate` and
-    /// > `AVSpeechUtteranceMaximumSpeechRate`. Lower values correspond to slower speech, and higher values correspond to
-    /// > faster speech. The default value is `AVSpeechUtteranceDefaultSpeechRate`.
-    /// > https://developer.apple.com/documentation/avfaudio/avspeechutterance/1619708-rate
-    private let avRateRange =
-        Double(AVSpeechUtteranceMinimumSpeechRate)...Double(AVSpeechUtteranceMaximumSpeechRate)
-
-    public lazy var defaultPitch: Double? =
-        avPitchRange.percentageForValue(1.0)
-
-    /// Range of valid values for an AVUtterance pitch.
-    ///
-    /// > Before enqueing the utterance, set this property to a value within the range of 0.5 for lower pitch to 2.0 for
-    /// > higher pitch. The default value is 1.0.
-    /// > https://developer.apple.com/documentation/avfaudio/avspeechutterance/1619683-pitchmultiplier
-    private let avPitchRange = 0.5...2.0
-
-    public weak var delegate: TTSEngineDelegate?
-
-    private let synthesizer = AVSpeechSynthesizer()
-
-    public override init() {
-        super.init()
-        synthesizer.delegate = self
-    }
-
-    public func speak(_ utterance: TTSUtterance) {
-        synthesizer.stopSpeaking(at: .immediate)
-        synthesizer.speak(avUtterance(from: utterance))
-    }
-
-    public func stop() {
-        synthesizer.stopSpeaking(at: .immediate)
-    }
-
-    private func avUtterance(from utterance: TTSUtterance) -> AVSpeechUtterance {
-        let avUtterance = AVUtterance(utterance: utterance)
-        if let rate = utterance.rate ?? defaultRate {
-            avUtterance.rate = Float(avRateRange.valueForPercentage(rate))
-        }
-        if let pitch = utterance.pitch ?? defaultPitch {
-            avUtterance.pitchMultiplier = Float(avPitchRange.valueForPercentage(pitch))
-        }
-        avUtterance.postUtteranceDelay = utterance.postDelay
-        avUtterance.voice = AVSpeechSynthesisVoice(language: utterance.language)
-        return avUtterance
-    }
-
-    public func speechSynthesizer(_ synthesizer: AVSpeechSynthesizer, didFinish utterance: AVSpeechUtterance) {
-        guard let utterance = utterance as? AVUtterance else {
-            return
-        }
-        delegate?.ttsEngine(self, didFinish: utterance.utterance)
-    }
-
-    public func speechSynthesizer(_ synthesizer: AVSpeechSynthesizer, willSpeakRangeOfSpeechString characterRange: NSRange, utterance avUtterance: AVSpeechUtterance) {
-        guard
-            let delegate = delegate,
-            let range = Range(characterRange)
-        else {
-            return
-        }
-//        controller?.notifySpeakingRange(range)
-    }
-
-    private class AVUtterance: AVSpeechUtterance {
-        let utterance: TTSUtterance
-
-        init(utterance: TTSUtterance) {
-            self.utterance = utterance
-            super.init(string: utterance.text)
-        }
-
-        required init?(coder: NSCoder) {
-            fatalError("Not supported")
         }
     }
 }
